backend/requirements.txt
fastapi
uvicorn
huggingface-hub==0.29.0
transformers==4.49.0
python-dotenv